moved override to command_line
