+++ src/input/mod.rs
@@ -1,17 +1,13 @@
 //! Input processing functionality for Sprocket.
 
 pub mod yaml;
-pub mod overrides;
-pub mod overrides_errors;
-#[cfg(test)]
-mod overrides_tests;
+pub mod command_line;
 
 // Re-export the main functionality from the yaml module
 pub use yaml::InputFormat;
 pub use yaml::get_json_file_path;
 pub use yaml::get_json_string;
 pub use yaml::parse_input_file;
-pub use overrides::InputOverride;
-pub use overrides::OverrideValue;
-pub use overrides::apply_overrides;
-pub use overrides_errors::*;
+
+// Re-export command line functionality
+pub use command_line::{CommandLineInput, InputValue, apply_inputs};
